+++ programs/drift/src/controller/orders/tests.rs
@@ -4014,9 +4014,9 @@ pub mod fill_order {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: oracle_price.twap as i64,
-                    last_oracle_price_twap_5min: oracle_price.twap as i64,
-                    last_oracle_price: oracle_price.agg.price as i64,
+                    last_oracle_price_twap: oracle_price.twap,
+                    last_oracle_price_twap_5min: oracle_price.twap,
+                    last_oracle_price: oracle_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -4218,9 +4218,9 @@ pub mod fill_order {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: oracle_price.twap as i64,
-                    last_oracle_price_twap_5min: oracle_price.twap as i64,
-                    last_oracle_price: oracle_price.agg.price as i64,
+                    last_oracle_price_twap: oracle_price.twap,
+                    last_oracle_price_twap_5min: oracle_price.twap,
+                    last_oracle_price: oracle_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -4557,9 +4557,9 @@ pub mod fill_order {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: oracle_price.twap as i64,
-                    last_oracle_price_twap_5min: oracle_price.twap as i64,
-                    last_oracle_price: oracle_price.agg.price as i64,
+                    last_oracle_price_twap: oracle_price.twap,
+                    last_oracle_price_twap_5min: oracle_price.twap,
+                    last_oracle_price: oracle_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -7824,9 +7824,9 @@ pub mod force_cancel_orders {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: oracle_price.twap as i64,
-                    last_oracle_price_twap_5min: oracle_price.twap as i64,
-                    last_oracle_price: oracle_price.agg.price as i64,
+                    last_oracle_price_twap: oracle_price.twap,
+                    last_oracle_price_twap_5min: oracle_price.twap,
+                    last_oracle_price: oracle_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -8120,9 +8120,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -8314,9 +8314,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -8509,9 +8509,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -8690,9 +8690,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -8945,9 +8945,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
@@ -9142,9 +9142,9 @@ pub mod get_maker_orders_info {
                 long_spread: 0,
                 short_spread: 0,
                 historical_oracle_data: HistoricalOracleData {
-                    last_oracle_price_twap: pyth_price.twap as i64,
-                    last_oracle_price_twap_5min: pyth_price.twap as i64,
-                    last_oracle_price: pyth_price.agg.price as i64,
+                    last_oracle_price_twap: pyth_price.twap,
+                    last_oracle_price_twap_5min: pyth_price.twap,
+                    last_oracle_price: pyth_price.agg.price,
                     ..HistoricalOracleData::default()
                 },
                 ..AMM::default()
