+++ programs/drift/src/math/orders/tests.rs
@@ -204,7 +204,7 @@ mod order_breaches_oracle_price_limits {
     #[test]
     fn bid_does_not_breach() {
         let _market = PerpMarket {
-            margin_ratio_initial: (MARGIN_PRECISION / 10) as u32, // 10x
+            margin_ratio_initial: (MARGIN_PRECISION / 10), // 10x
             ..PerpMarket::default()
         };
 
@@ -236,7 +236,7 @@ mod order_breaches_oracle_price_limits {
     #[test]
     fn bid_does_not_breach_4_99_percent_move() {
         let _market = PerpMarket {
-            margin_ratio_initial: (MARGIN_PRECISION / 10) as u32, // 10x
+            margin_ratio_initial: (MARGIN_PRECISION / 10), // 10x
             ..PerpMarket::default()
         };
 
@@ -268,8 +268,8 @@ mod order_breaches_oracle_price_limits {
     #[test]
     fn bid_breaches() {
         let _market = PerpMarket {
-            margin_ratio_initial: (MARGIN_PRECISION / 10) as u32, // 10x
-            margin_ratio_maintenance: (MARGIN_PRECISION / 20) as u32, // 20x
+            margin_ratio_initial: (MARGIN_PRECISION / 10), // 10x
+            margin_ratio_maintenance: (MARGIN_PRECISION / 20), // 20x
             ..PerpMarket::default()
         };
 
@@ -302,8 +302,8 @@ mod order_breaches_oracle_price_limits {
     #[test]
     fn ask_does_not_breach() {
         let _market = PerpMarket {
-            margin_ratio_initial: (MARGIN_PRECISION / 10) as u32, // 10x
-            margin_ratio_maintenance: (MARGIN_PRECISION / 20) as u32, // 20x
+            margin_ratio_initial: (MARGIN_PRECISION / 10), // 10x
+            margin_ratio_maintenance: (MARGIN_PRECISION / 20), // 20x
             ..PerpMarket::default()
         };
 
@@ -336,8 +336,8 @@ mod order_breaches_oracle_price_limits {
     #[test]
     fn ask_does_not_breach_4_99_percent_move() {
         let _market = PerpMarket {
-            margin_ratio_initial: (MARGIN_PRECISION / 10) as u32, // 10x
-            margin_ratio_maintenance: (MARGIN_PRECISION / 20) as u32, // 20x
+            margin_ratio_initial: (MARGIN_PRECISION / 10), // 10x
+            margin_ratio_maintenance: (MARGIN_PRECISION / 20), // 20x
             ..PerpMarket::default()
         };
 
@@ -370,8 +370,8 @@ mod order_breaches_oracle_price_limits {
     #[test]
     fn ask_breaches() {
         let _market = PerpMarket {
-            margin_ratio_initial: (MARGIN_PRECISION / 10) as u32, // 10x
-            margin_ratio_maintenance: (MARGIN_PRECISION / 20) as u32, // 20x
+            margin_ratio_initial: (MARGIN_PRECISION / 10), // 10x
+            margin_ratio_maintenance: (MARGIN_PRECISION / 20), // 20x
             ..PerpMarket::default()
         };
